Add checker for OOOA
An Out-Of-Order Alternation is a subtle bug.  When ordering an alternation
lexicographically, you commonly end up with something like (else|elseif).
For simple regexes like this, Python picks the first choice that matches, not
the longest choice (as Perl would).  You can solve this by putting an anchor
after the alternations, i.e. (else|elseif)\b or simply (elseif|else).

This checker is a quick and dirty version, and doesn't handle the full
complexity of the issue, but finds the obvious ones much better than my eyes
can.