Quickfix for simplify_charclass: do not consider UNICODE regexes
In re.UNICODE mode (which is the default on Py3), the builtin
charclasses \wWsSdD match more than just ASCII.  The logic can't
handle this currently, so disable the checker.

Also, the way negated classes are handled (by building the full
complement set) will have to be changed to avoid storing millions
of Unicode characters over and over.